+++ webpack.docs.config.js
@@ -1,10 +1,7 @@
 const path = require('path');
-const autoprefixer = require('autoprefixer');
-const precss = require('precss');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const webpack = require('webpack');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const flexbugs = require('postcss-flexbugs-fixes');
 const merge = require('webpack-merge');
 const getBaseConfiguration = require('./webpack/base.config.js');
 
@@ -33,78 +30,18 @@ const plugins = [
     filename: 'index.html',
     template: 'examples/index.html',
   }),
-  new webpack.DefinePlugin({
-    'process.env': {
-      NODE_ENV: JSON.stringify('production'),
-    },
-  }),
   new ExtractTextPlugin({
     filename: 'styles/[name].[contenthash].css',
     allChunks: true,
   }),
-  /* new webpack.optimize.UglifyJsPlugin({
-    compress: {
-      warnings: true,
-    },
-    output: {
-      comments: false,
-    },
-    mangle: true,
-  }), */
   new webpack.ProvidePlugin({
     $: 'jquery',
     jQuery: 'jquery',
   }),
 ];
 
-const rules = [
-  {
-    test: /\.ejs$/,
-    use: [
-      'ejs-loader?variable=data',
-    ],
-  },
-  {
-    test: /\.scss$/,
-    use: [
-      'style-loader',
-      'css-loader',
-      {
-        loader: 'postcss-loader',
-        options: {
-          plugins: () => [flexbugs, precss, autoprefixer],
-        },
-      },
-      'sass-loader',
-    ],
-  },
-  {
-    test: /\.ico$/,
-    use: [
-      'file-loader?name=[name].[ext]',
-    ],
-    include: /images/,
-  },
-  {
-    test: /\.css$/,
-    use: [
-      'style-loader',
-      'css-loader',
-      {
-        loader: 'postcss-loader',
-        options: {
-          plugins: () => [flexbugs, precss, autoprefixer],
-        },
-      },
-    ],
-  },
-];
-
 const config = merge(getBaseConfiguration(params), {
   plugins,
-  module: {
-    rules,
-  },
 });
 
 module.exports = config;