+++ webpack/base.config.js
@@ -1,11 +1,26 @@
+/* eslint-disable global-require, no-console */
 const path = require('path');
-const webpack = require('webpack');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
+const merge = require('webpack-merge');
+const autoprefixer = require('autoprefixer');
+const precss = require('precss');
+const flexbugs = require('postcss-flexbugs-fixes');
+const colors = require('colors/safe');
+const utils = require('./utils.js');
+
+const isProduction = utils.isProduction();
+
+let getBaseEnvConfiguration;
+
+if (isProduction) {
+  getBaseEnvConfiguration = require('./base.production.config.js');
+} else {
+  getBaseEnvConfiguration = require('./base.development.config.js');
+}
 
 function getBaseConfiguration(config) {
-  return {
+  console.log(`${colors.green.underline('Building:')} ${colors.yellow(`${config.buildPath}/${config.output.filename}`)} ${colors.grey(isProduction ? '[production]' : '[development]')}`);
+  return merge(getBaseEnvConfiguration(config), {
     entry: config.entry,
-    devtool: 'source-map',
     output: config.output,
     module: {
       rules: [
@@ -20,6 +35,48 @@ function getBaseConfiguration(config) {
           test: /\.svg$/,
           use: ['babel-loader', 'react-svg-loader'],
         },
+        {
+          test: /\.ejs$/,
+          use: [
+            'ejs-loader?variable=data',
+          ],
+        },
+        {
+          test: /\.scss$/,
+          use: [
+            'style-loader',
+            'css-loader',
+            {
+              loader: 'postcss-loader',
+              options: {
+                plugins: () => [flexbugs, precss, autoprefixer],
+                minimize: isProduction,
+              },
+            },
+            'sass-loader',
+          ],
+        },
+        {
+          test: /\.ico$/,
+          use: [
+            'file-loader?name=[name].[ext]',
+          ],
+          include: /images/,
+        },
+        {
+          test: /\.css$/,
+          use: [
+            'style-loader',
+            'css-loader',
+            {
+              loader: 'postcss-loader',
+              options: {
+                plugins: () => [flexbugs, precss, autoprefixer],
+                minimize: isProduction,
+              },
+            },
+          ],
+        },
       ],
     },
     resolve: {
@@ -29,15 +86,7 @@ function getBaseConfiguration(config) {
       ],
       extensions: ['.js', '.jsx'],
     },
-    plugins: [
-      new CleanWebpackPlugin([config.buildPath], {
-        root: config.root,
-        verbose: false,
-        dry: false,
-      }),
-      new webpack.optimize.ModuleConcatenationPlugin(),
-    ],
-  };
+  });
 }
 
 module.exports = getBaseConfiguration;