+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Include both development and production builds
 
 ## 2.1.2
 * Change to development build to avoid warnings